requirements.txt
baserowapi==0.1.0b4
requests==2.32.3
pytz==2024.2
pyyaml==6.0.2
python-dotenv=1.0.1